Refactor canvas
Decouple canvas implementation into a separate component, this makes it
easier to switch to other canvas implementations like PathKit or
CanvasKit later
