+++ public/index.html
@@ -23,7 +23,7 @@
   </head>
   <body>
     <main-application></main-application>
-    <script type='module' src="./dist/main-application.js"></script>
+    <script type="module" src="./dist/main-application.js"></script>
     <noscript>
       This page requires JavaScript scripting, please turn it on.
     </noscript>
