+++ src/main-application.js
@@ -4,8 +4,9 @@ import '@material/mwc-drawer';
 import '@material/mwc-icon-button';
 import '@material/mwc-snackbar';
 import '@material/mwc-top-app-bar';
-import './toolbar.js';
 import './info-panel.js';
+import './js-canvas.js';
+import './toolbar.js';
 
 export class MainApplication extends LitElement {
   static styles = css`
@@ -43,26 +44,17 @@ export class MainApplication extends LitElement {
       font-weight: bold;
     }
 
-    tiny-toolbar{
+    tiny-toolbar {
       width: 20vw;
       height: 100vh;
     }
 
-    #canvas {
+    main-canvas {
       position: absolute;
+      top: 0px;
+      left: 0px;
       width: 100vw;
       height: 100vh;
-      user-select: none;
-      touch-action: none;
-      z-index: 1;
-    }
-
-    #prediction-canvas {
-      position: absolute;
-      pointer-events: none;
-      user-select: none;
-      touch-action: none;
-      z-index: 2;
     }
 
     info-panel {
@@ -81,11 +73,6 @@ export class MainApplication extends LitElement {
   `;
 
   _rafId = null;
-  _pointerDown = false;
-  _pointerMoved = false;
-  _currentColor = '#000000';
-  _points = [];
-  _predicted_points = [];
 
   // for overlay drag
   _dragActive = false;
@@ -103,16 +90,6 @@ export class MainApplication extends LitElement {
         container.addEventListener('MDCTopAppBar:nav', () => {
             this._drawer.open = !this._drawer.open;
         });
-    }
-
-    this._canvas = this.shadowRoot.querySelector('#canvas');
-    if (this._canvas && this._canvas.getContext)
-      this._context = this._canvas.getContext('2d');
-
-    // Check that we have a valid context to draw on/with before adding event handlers
-    if (!this._context) {
-      console.error('Your browser doesn\'t support canvas, this demo won\'t work');
-      return;
     }
 
     this._snackbar = this.shadowRoot.querySelector('#snackbar');
@@ -141,48 +118,22 @@ export class MainApplication extends LitElement {
       });
     }
 
+    this._mainCanvas = this.shadowRoot.querySelector('#main-canvas');
+    this._mainCanvas.app = this;
     this._infoButton = this.shadowRoot.querySelector('#info-button');
     this._infoButton.onpointerdown = this._toggleInfoPanel.bind(this);
     this._clearButton = this.shadowRoot.querySelector('#clear-button');
     this._clearButton.onpointerdown = this._clearCanvas.bind(this);
-    this._canvas.onpointerdown = this._onPointerDown.bind(this);
-    this._canvas.onpointermove = this._onPointerMove.bind(this);
-    this._canvas.onpointerup = this._onPointerUp.bind(this);
-    const style = window.getComputedStyle(this._canvas);
-    this._canvas.width  = parseInt(style.width, 10);
-    this._canvas.height =  parseInt(style.height, 10);
-    this._predictionCanvas = this.shadowRoot.querySelector('#prediction-canvas');
-    this._predictionCanvas.style.left = style.left + 'px';
-    this._predictionCanvas.style.top = style.top + 'px';
-    this._predictionCanvas.width = this._canvas.width;
-    this._predictionCanvas.height = this._canvas.height;
-    this._canvas.style.top = this._predictionCanvas.style.top = 0;
-    this._canvas.style.left = this._predictionCanvas.style.left = 0;
-    this._predictionCanvasContext = this._predictionCanvas.getContext('2d');
-    this._context.lineCap = this._predictionCanvasContext.lineCap = 'round';
-    this._context.lineJoin = this._predictionCanvasContext.lineJoin = 'round';
-    this._context.shadowBlur = this._predictionCanvasContext.shadowBlur = 2;
     this._infoPanel = this.shadowRoot.querySelector('#info-panel');
     this._infoPanel.style.visibility = 'hidden';
     this._infoPanel.onpointerdown = this._onDragStart.bind(this);
     this._infoPanel.onpointermove = this._onDrag.bind(this);
     this._infoPanel.onpointerup = this._onDragEnd.bind(this);
-    window.addEventListener('resize', this._onResize);
     console.log(window.navigator.usi);
   }
 
   constructor() {
     super();
-    this._drawWithPreferredColor = false;
-    this._drawWithPressure = false;
-    this._drawPredictedEvents = false;
-    this._highlightPredictedEvents = false;
-    this._drawCoalescedEvents = false;
-    this._drawPointsOnly = false;
-    this._currentLineWidth = 8;
-    // By default, the number of points slider is set to 2, because the first 2
-    // predictions seems to be a good number, the other predictions are very far off.
-    this._numOfPredictionPoints = 2;
   }
 
   _showSnackbar() {
@@ -194,72 +145,7 @@ export class MainApplication extends LitElement {
   }
 
   _clearCanvas = async (event) => {
-    this._context.clearRect(0, 0, this._context.canvas.width, this._context.canvas.height);
-    this._predictionCanvasContext.clearRect(0, 0,
-      this._predictionCanvasContext.canvas.width, this._predictionCanvasContext.canvas.height);
-  }
-
-  _onPointerDown = async (event) => {
-    this._pointerDown = true;
-    this._pointerId = event.pointerId;
-    this._canvas.setPointerCapture(this._pointerId);
-    this._points.push(event);
-    this._rafId = window.requestAnimationFrame(this._onAnimationFrame.bind(this));
-    event.preventDefault();
-  }
-
-  _onPointerMove = async (event) => {
-    if (event.clientY < 0 || event.clientX < 0 ||
-      event.clientX > window.innerWidth || event.clientY > window.innerHeight) {
-        this._pointerDown = false;
-        return;
-    }
-
-    if(this._pointerDown) {
-      this._pointerMoved = true;
-      if (event.getCoalescedEvents && this._drawCoalescedEvents) {
-        if (event.getCoalescedEvents().length > 0) {
-          for (let e of event.getCoalescedEvents())
-            this._points.push(e);
-        } else {
-          this._points.push(event);
-        }
-      } else {
-        this._points.push(event);
-      }
-
-      if (this._drawPredictedEvents && event.getPredictedEvents) {
-        for (let e of event.getPredictedEvents()) {
-          this._predicted_points.push(e);
-        }
-
-        // number of points from slider should be between 1 - 10
-        if (this._numOfPredictionPoints > 0 && this._numOfPredictionPoints <= 10)
-          this._predicted_points = this._predicted_points.slice(0, this._numOfPredictionPoints);
-      }
-
-      this._rafId = window.requestAnimationFrame(this._onAnimationFrame.bind(this));
-      event.preventDefault();
-    }
-  }
-
-  _onPointerUp = async (event) => {
-    if (this._rafId) {
-      window.cancelAnimationFrame(this._rafId);
-      this._rafId = null;
-    }
-
-    if (this._drawPredictedEvents)
-      this._predictionCanvasContext.clearRect(0, 0,
-        this._predictionCanvasContext.canvas.width, this._predictionCanvasContext.canvas.height);
-
-    if (this.pointerMoved)
-      this._pointerMoved = false;
-    this._pointerDown = false;
-    this._canvas.releasePointerCapture(this._pointerId);
-    this._predicted_points = [];
-    this._points = [];
-    this._updateInfoPanel(event);
+    this._mainCanvas._clearCanvas();
   }
 
   // for overlay drag
@@ -296,116 +182,8 @@ export class MainApplication extends LitElement {
   }
 
   _onAnimationFrame() {
-    if(this._pointerDown && this._points.length > 0) {
-      if (this._drawPointsOnly)
-        this._drawPoints(this._context);
-      else
-        this._drawStroke(this._context);
-
-      if (this._drawPredictedEvents && this._predicted_points.length > 0) {
-        // This will clear the canvas (which include the previous predictions).
-        this._predictionCanvasContext.clearRect(0, 0,
-          this._predictionCanvasContext.canvas.width, this._predictionCanvasContext.canvas.height);
-        this._strokePredictedEvents(this._predictionCanvasContext);
-      }
-
-      // Drop all previous coalesced pointer events except for the last one
-      // which is used for the next start position for the stroke.  If
-      // coalesced events were not used, then the last point will always be
-      // the current x y position of pointerMove event.
-      this._points.splice(0, this._points.length-1);
-      this._predicted_points = [];
-      this._updateInfoPanel(this._points[0]);
-    } else if (this._dragActive) {
+    if (this._dragActive) {
       this._infoPanel.style.transform = 'translate3d(' + this._currentX + 'px, ' + this._currentY + 'px, 0)';
-    }
-  }
-
-  _drawStroke(context) {
-    if (this._points.length < 2) {
-      context.beginPath();
-      context.fillStyle = this._getCurrentColor(this._points[0]);
-      let radius;
-      if (this._drawWithPressure)
-        radius = this._currentLineWidth * this._points[0].pressure;
-      else
-        radius = this._currentLineWidth / 2;
-      context.arc(this._points[0].x, this._points[0].y, radius, 0, Math.PI * 2, true);
-      context.fill();
-      return;
-    }
-
-    let i;
-    for (i = 0; i < this._points.length-1; i++) {
-      let startWidth, endWidth;
-      // Varying brush size based on pressure, convert from pressure range of 0 to 1
-      // to a scale factor of 0 to 2
-      if (this._drawWithPressure) {
-        startWidth = this._currentLineWidth * this._points[i].pressure * 2;
-        endWidth = this._currentLineWidth * this._points[i+1].pressure * 2;
-      } else {
-        startWidth = endWidth = this._currentLineWidth;
-      }
-      let path = this._createPath(this._points[i].x, this._points[i].y, this._points[i+1].x, this._points[i+1].y, startWidth, endWidth);
-      context.fillStyle = this._getCurrentColor(this._points[i]);
-      context.fill(path);
-    }
-  }
-
-  _drawPoints(context) {
-    if (this._points.length < 2) {
-      context.beginPath();
-      context.fillStyle = this._getCurrentColor(this._points[0]);
-      context.arc(this._points[0].x,  this._points[0].y, 3, 0, Math.PI * 2, true);
-      context.fill();
-      return;
-    }
-
-    let i;
-    for (i = 1; i < this._points.length-1; i++) {
-      context.beginPath();
-      context.fillStyle = '#FF0000';
-      context.arc(this._points[i].x, this._points[i].y, 2, 0, Math.PI * 2, true);
-      context.fill();
-    }
-
-    context.beginPath();
-    context.fillStyle = this._getCurrentColor(this._points[i]);
-    context.arc(this._points[i].x, this._points[i].y, 3, 0, Math.PI * 2, true);
-    context.fill();
-  }
-
-  _createPath(x1, y1, x2, y2, startWidth, endWidth) {
-    const vectorX = x2 - x1,
-          vectorY = y2 - y1;
-    const vectorAngle = Math.atan2(vectorY, vectorX) + Math.PI / 2;
-    const path = new Path2D();
-
-    path.arc(x1, y1, startWidth / 2, vectorAngle, vectorAngle + Math.PI);
-    path.arc(x2, y2, endWidth / 2, vectorAngle + Math.PI, vectorAngle);
-    path.closePath();
-    return path;
-  }
-
-  _strokePredictedEvents(context) {
-    if (this._points.length > 0) {
-      // Varying brush size based on pressure, convert from pressure range of 0 to 1
-      // to a scale factor of 0 to 2
-      let lastPoint = this._points[this._points.length-1];
-      if (this._drawWithPressure)
-        context.lineWidth = this._currentLineWidth * lastPoint.pressure * 2;
-      else
-        context.lineWidth = this._currentLineWidth;
-      context.beginPath();
-      context.moveTo(lastPoint.x, lastPoint.y);
-      if (this._highlightPredictedEvents)
-        context.strokeStyle = 'red';
-      else
-        context.strokeStyle = this._getCurrentColor(lastPoint);
-      for (let p of this._predicted_points) {
-        context.lineTo(p.x, p.y);
-      }
-      context.stroke();
     }
   }
 
@@ -431,54 +209,41 @@ export class MainApplication extends LitElement {
     return Math.round(value * factor) / factor;
   }
 
-  _getCurrentColor(event) {
-    if (event.preferredColor && this._drawWithPreferredColor)
-      return event.preferredColor;
-    else
-      return this._currentColor;
-  }
-
   _colorChanged(event) {
-    this._currentColor = event.detail.color;
-    this._drawWithPreferredColor = false;
+    this._mainCanvas.currentColor = event.detail.color;
+    this._mainCanvas.drawWithPreferredColor = false;
   }
 
   _lineWidthChanged(event) {
-    this._currentLineWidth = event.detail.lineWidth;
+    this._mainCanvas.currentLineWidth = event.detail.lineWidth;
   }
 
   _drawWithPreferredColorChanged(event) {
-    this._drawWithPreferredColor = event.detail.drawWithPreferredColor;
+    this._mainCanvas.drawWithPreferredColor = event.detail.drawWithPreferredColor;
   }
 
   _pressureEventsEnabledChanged(event) {
-    this._drawWithPressure = event.detail.pressureEventsEnabled;
+    this._mainCanvas.drawWithPressure = event.detail.pressureEventsEnabled;
   }
 
   _predictedEventsEnabledChanged(event) {
-    this._drawPredictedEvents = event.detail.predictedEventsEnabled;
+    this._mainCanvas.drawPredictedEvents = event.detail.predictedEventsEnabled;
   }
 
   _predictedEventsHighlightEnabledChanged(event) {
-    this._highlightPredictedEvents = event.detail.predictedEventsHighlightEnabled;
+    this._mainCanvas.highlightPredictedEvents = event.detail.predictedEventsHighlightEnabled;
   }
 
   _numOfPredictionPointsChanged(event) {
-    this._numOfPredictionPoints = event.detail.numOfPredictionPoints;
+    this._mainCanvas.numOfPredictionPoints = event.detail.numOfPredictionPoints;
   }
 
   _coalescedEventsEnabledChanged(event) {
-    this._drawCoalescedEvents = event.detail.coalescedEventsEnabled;
+    this._mainCanvas.drawCoalescedEvents = event.detail.coalescedEventsEnabled;
   }
 
   _drawPointsOnlyEnabledChanged(event) {
-    this._drawPointsOnly = event.detail.drawPointsOnlyEnabled;
-  }
-
-  _onResize = async (event) => {
-    const style = window.getComputedStyle(this._canvas);
-    this._canvas.width  = this._predictionCanvas.width = parseInt(style.width, 10);
-    this._canvas.height =  this._predictionCanvas.height = parseInt(style.height, 10);
+    this._mainCanvas.drawPointsOnly = event.detail.drawPointsOnlyEnabled;
   }
 
   render() {
@@ -504,8 +269,7 @@ export class MainApplication extends LitElement {
           <mwc-icon-button slot="actionItems" id="info-button" icon="info"></mwc-icon-button>
           <mwc-icon-button slot="actionItems" id="clear-button" icon="clear"></mwc-icon-button>
         </mwc-top-app-bar>
-        <canvas id="canvas"></canvas>
-        <canvas id="prediction-canvas"></canvas>
+        <main-canvas id="main-canvas"></main-canvas>
         <info-panel id="info-panel"></info-panel>
       </div>
     </mwc-drawer>
